requirements.txt
aioredis==1.3.1
anyio==3.3.4
asgiref==3.4.1
async-timeout==4.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.3
fastapi==0.70.0
fastapi-limiter==0.1.4
h11==0.12.0
hiredis==2.0.0
httptools==0.3.0
idna==3.3
motor==2.5.1
pydantic==1.8.2
pymongo==3.12.3
python-dotenv==0.19.2
PyYAML==6.0
requests==2.27.1
sniffio==1.2.0
starlette==0.16.0
typing-extensions==3.10.0.2
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.2

PyJWT~=2.3.0